+++ src/main/java/ru/whoisamyy/api/gd/objects/Comment.java
@@ -123,7 +123,7 @@ public class Comment extends GDObject {
         return sb.substring(0, sb.length()-1);
     }
 
-    public byte upload() {
+    public int upload() {
         if (!Account.getAccountsHashtable().containsKey(accountID) && isAcc) return -1;
         if (!Level.getLevelsHashtable().containsKey(levelID) && !isAcc) return -1;
 
@@ -146,7 +146,7 @@ public class Comment extends GDObject {
         }
     }
 
-    public byte delete() {
+    public int delete() {
         try(PreparedStatement ps = conn.prepareStatement("DELETE FROM comments WHERE ID = ? LIMIT 1")) {
             ps.setInt(1, getID());
 
@@ -160,7 +160,7 @@ public class Comment extends GDObject {
     }
 
     @Deprecated
-    public byte uploadAccComment() {
+    public int uploadAccComment() {
         if (!isAcc) return -1;
         if (!Account.getAccountsHashtable().containsKey(accountID)) return -1;
         if (!Level.getLevelsHashtable().containsKey(levelID)) return -1;
